Add switch network
